requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pymongo==4.8.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pydantic==2.9.0
pydantic-settings==2.5.2
playwright==1.48.0
requests==2.31.0
pytrends==4.9.2
